+++ appboy-component/samples/xamarin-forms/IAppboyPlatformDelegate.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace TestAppXamarinForms
+{
+  public interface IAppboyPlatformDelegate
+  {
+    void LogCustomEvent(String eventName);
+  }
+}
+