+++ appboy-component/samples/xamarin-forms/MenuPage.xaml.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+
+using Xamarin.Forms;
+
+namespace TestAppXamarinForms
+{
+  public partial class MenuPage : ContentPage
+  {
+    public MenuPage ()
+    {
+      InitializeComponent();
+    }
+
+    void OnLogCustomEvent(object sender, EventArgs args)
+    {
+      DependencyService.Get<IAppboyPlatformDelegate>().LogCustomEvent("FormsSampleEvent");
+    }
+  }
+}
+