+++ appboy-component/src/ios-unified/StructsAndEnums.cs
@@ -2,62 +2,66 @@
 
 namespace AppboyPlatformXamariniOSBinding
 {
-  public enum ABKRequestProcessingPolicy : long /* nint */ {
+  public enum ABKRequestProcessingPolicy /* nint */ {
     ABKAutomaticRequestProcessing,
     ABKAutomaticRequestProcessingExceptForDataFlush,
     ABKManualRequestProcessing
   }
 
-  public enum ABKSocialAccountAcquisitionPolicy : long /* nint */ {
+  public enum ABKSocialAccountAcquisitionPolicy /* nint */ {
     ABKAutomaticSocialAccountAcquisition,
     ABKAutomaticSocialAccountAcquisitionWithIdentifierOnly,
     ABKManualSocialAccountAcquisition
   }
 
-  public enum ABKSocialNetwork : ulong /* nuint */ {
+  public enum ABKSocialNetwork /* nuint */ {
     Facebook = 1 << 0,
     Twitter = 1 << 1
   }
 
-  public enum ABKUserGenderType : long /* nint */ {
+  public enum ABKUserGenderType /* nint */ {
     Male,
     Female
   }
 
-  public enum ABKNotificationSubscriptionType : long /* nint */ {
+  public enum ABKNotificationSubscriptionType /* nint */ {
     ABKOptedIn,
     ABKSubscribed,
     ABKUnsubscribed
   }
 
-  public enum ABKSlideupAnchor : long /* nint */ {
-    ABKSlideupFromTop,
-    ABKSlideupFromBottom
+  public enum ABKInAppMessageClickActionType
+  {
+    DisplayNewsFeed,
+    RedirectToURI,
+    NoneClickAction
   }
 
-  public enum ABKSlideupDismissType : long /* nint */ {
+  public enum ABKInAppMessageDismissType
+  {
     Automatically,
-    BySwipe
+    Manually
   }
 
-  public enum ABKSlideupClickActionType : long /* nint */ {
-    ABKSlideupDisplayNewsFeed,
-    ABKSlideupRedirectToURI,
-    ABKSlideupNoneClickAction
+  public enum ABKInAppMessageDisplayChoice
+  {
+    playInAppMessageNow,
+    playInAppMessageLater,
+    cardInAppMessage
   }
-
-  public enum ABKSlideupDisplayChoice : long /* nint */ {
-    ABKDisplaySlideupNow,
-    ABKDisplaySlideupLater,
-    ABKDiscardSlideup
-  }
-  public enum ABKCardCategory : ulong /* nuint */ {
+  public enum ABKCardCategory /* nuint */ {
     NoCategory = 1 << 0,
     News = 1 << 1,
     Advertising = 1 << 2,
     Announcements = 1 << 3,
     Social = 1 << 4,
     All = 1 << 0 | 1 << 1 | 1 << 2 | 1 << 3 | 1 << 4
+  }
+      
+  public enum ABKInAppMessageSlideupAnchor
+  {
+    Top,
+    Bottom
   }
 }