+++ appboy-component/src/ios-unified/libAppboyKit.linkwith.cs
@@ -1,4 +1,4 @@
 using System;
 using ObjCRuntime;
 
-[assembly: LinkWith ("libAppboyKit.a", LinkTarget.ArmV7 | LinkTarget.ArmV7s | LinkTarget.Arm64 | LinkTarget.Simulator, Frameworks="SystemConfiguration QuartzCore CoreImage CoreTelephony Social Twitter Accounts AdSupport StoreKit", ForceLoad = true, SmartLink = true)]
+[assembly: LinkWith ("libAppboyKit.a", LinkTarget.ArmV7 | LinkTarget.ArmV7s | LinkTarget.Simulator | LinkTarget.Simulator64 | LinkTarget.Arm64, Frameworks="SystemConfiguration QuartzCore CoreImage CoreTelephony Social Accounts AdSupport StoreKit CoreText", SmartLink = true, ForceLoad = true)]